backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
bcrypt>=4.0.1
python-multipart>=0.0.5
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
requests>=2.31.0
# AI/ML
google-genai>=0.3.0
groq>=0.4.0